Add basic yaml generating
This is just preliminary version of our yaml scheme, it's missing "state"
of the app (need to check, currently using etc.), it will be changed
when final version of discussion and yaml formats will be established.

Generated yaml data is only printed, but will be also dumped to the file
if filename is provided.
Created App class for code readability.